Project Parked in a broken state.
syscfg generation is not occurring correctly in this project, includes were pointing to a one-time-generated version of the syscfg files.

this means includes which rely on board definitions must be modified.

code otherwise is an import from a working project.  Carefully fix syscfg to generate files once more for this project.